Version2.0
Player controls have finally been fine tuned, the enemy AI is now able to detect both pits and walls and react accordingly to them. All that is left is to create death() for both player and enemy. Also forcing them to reset the scene once the player has completed the level.